# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Booking API documentation",
    "description": "L5 Swagger OpenApi description",
    "contact": {
      "email": "khujand.nanocoding@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001/api",
      "description": "Local API Server"
    }
  ],
  "paths": {
...
    "/company": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "Display all companies",
        "operationId": "companyAll",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The company`s name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "description": "The company`s description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "query",
            "description": "The company`s address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "in": "query",
            "description": "The company`s phone",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "The company`s statuses",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Create a company",
        "operationId": "companyPost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/company/{id}": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "Show company by id",
        "operationId": "showCompanyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The company`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Company"
        ],
        "summary": "Update company by id",
        "operationId": "updateCompanyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The company`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Company"
        ],
        "summary": "Delete company by id",
        "operationId": "deleteCompanyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The company`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/company/toggle/{id}": {
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Toggle status company by id",
        "operationId": "toggleStatusCompanyById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The company`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/table": {
      "get": {
        "tags": [
          "Table"
        ],
        "summary": "Display all tables",
        "operationId": "tableAll",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The table`s name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seat_count",
            "in": "query",
            "description": "The table`s seat count",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "The table`s seat type (TABLE = 0; CABIN = 1;)",
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "The table`s status",
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Table"
        ],
        "summary": "Create a table",
        "operationId": "tablePost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/table/{id}": {
      "get": {
        "tags": [
          "Table"
        ],
        "summary": "Show table by id",
        "operationId": "showTableById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The table`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Table"
        ],
        "summary": "Update table by id",
        "operationId": "updateTableById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The table`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Table"
        ],
        "summary": "Delete table by id",
        "operationId": "deleteTableById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The table`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/table/toggle/{id}": {
      "post": {
        "tags": [
          "Table"
        ],
        "summary": "Toggle status table by id",
        "operationId": "toggleStatusTableById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The table`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyModel": {
        "title": "Company model",
        "description": "Company model",
        "properties": {
          "id": {
            "title": "Id",
            "description": "Company`s id",
            "format": "string"
          },
          "name": {
            "title": "Name",
            "description": "Company`s name",
            "format": "string"
          },
          "description": {
            "title": "Description",
            "description": "Company`s description",
            "format": "string"
          },
          "image": {
            "title": "Image",
            "description": "Company`s image",
            "format": "string"
          },
          "preview": {
            "title": "Preview",
            "description": "Company`s preview",
            "format": "string"
          },
          "address": {
            "title": "Address",
            "description": "Company`s address",
            "format": "string"
          },
          "phone": {
            "title": "Phone",
            "description": "Company`s phone",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Company`s status",
            "format": "string"
          },
          "created_at": {
            "title": "Created at",
            "description": "Company`s created at",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated at",
            "description": "Company`s updated at",
            "format": "date-time"
          }
        },
        "type": "object"
      },
      "CompanyStoreModel": {
        "title": "Company model",
        "description": "Company model",
        "properties": {
          "name": {
            "title": "Name",
            "description": "Company`s name",
            "format": "string"
          },
          "description": {
            "title": "Description",
            "description": "Company`s description",
            "format": "string"
          },
          "image": {
            "title": "Image",
            "description": "Company`s image",
            "format": "string"
          },
          "address": {
            "title": "Address",
            "description": "Company`s address",
            "format": "string"
          },
          "phone": {
            "title": "Phone",
            "description": "Company`s phone",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Company`s status",
            "format": "string"
          }
        },
        "type": "object"
      },
...
      "TableModel": {
        "title": "Table model",
        "description": "Table model",
        "properties": {
          "id": {
            "title": "Id",
            "description": "Table`s id",
            "format": "string"
          },
          "name": {
            "title": "Name",
            "description": "Table`s name",
            "format": "string"
          },
          "seat_count": {
            "title": "Seat count",
            "description": "Table`s seat count",
            "format": "integer"
          },
          "type": {
            "title": "type",
            "description": "Table`s type (TABLE = '0'; CABIN = '1';)",
            "format": "string"
          },
          "company_id": {
            "title": "Company id",
            "description": "Table`s company id",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Table`s status",
            "format": "string"
          },
          "created_at": {
            "title": "Created at",
            "description": "Table`s created at",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated at",
            "description": "Table`s updated at",
            "format": "date-time"
          }
        },
        "type": "object"
      },
      "TableStoreModel": {
        "title": "Table store model",
        "description": "Table store model",
        "properties": {
          "name": {
            "title": "Name",
            "description": "Table`s name",
            "format": "string"
          },
          "seat_count": {
            "title": "Seat count",
            "description": "Table`s seat count",
            "format": "integer"
          },
          "type": {
            "title": "type",
            "description": "Table`s type (TABLE = '0'; CABIN = '1';)",
            "format": "string"
          },
          "company_id": {
            "title": "Company id",
            "description": "Table`s company id",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Table`s status",
            "format": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}